EES-1855 Move all pre-release components into their own directory
