+++ src/explore-education-statistics-admin/src/routes/routes.ts
@@ -7,8 +7,8 @@ import MethodologiesPage from '@admin/pages/methodology/MethodologiesPage';
 import MethodologyCreatePage from '@admin/pages/methodology/MethodologyCreatePage';
 import PublicationCreatePage from '@admin/pages/publication/PublicationCreatePage';
 import PublicationEditPage from '@admin/pages/publication/PublicationEditPage';
-import PreReleaseAccessListPage from '@admin/pages/release/PreReleaseAccessListPage';
-import PreReleasePage from '@admin/pages/release/PreReleasePage';
+import PreReleaseAccessListPage from '@admin/pages/release/pre-release/PreReleaseAccessListPage';
+import PreReleasePage from '@admin/pages/release/pre-release/PreReleasePage';
 import ReleaseCreatePage from '@admin/pages/release/ReleaseCreatePage';
 import ReleaseMetaGuidancePage from '@admin/pages/release/ReleaseMetaGuidancePage';
 import ReleasePageContainer from '@admin/pages/release/ReleasePageContainer';
